+++ include/glex/Sphere.h
@@ -3,10 +3,10 @@
 
 class Sphere {
     public:
-        Sphere(double radius);
+        Sphere(float radius);
         void draw();
     private:
-        double _radius;
+        float _radius;
         void _drawFace(int p_recurse, GLfloat* a, GLfloat* b, GLfloat* c);
         void _drawList();
         // GLint _list;
